A closeable* may be closed even if not in closeable map
A bug is an unintended behaviour. At first I didn't think deep enough
about closeable* and didn't imagine they could be used to mark a
closeable object not in the map, but internally referenced by a map
value. Now that works, too. It's no longer an omition but a feature.

Also, this commit clarifies how to deal with instances of Java
objects, and extends closeable-map* so that it may take a body that
eventually returns a map.